+++ src/lib-components/Flexible/MediaGallery.vue
@@ -1,14 +1,16 @@
 <template>
     <section class="media-gallery">
-        <h2 class="section-title">
-            {{ block.titleGeneral }}
-        </h2>
+        <div class="section-header">
+            <h2 class="section-title">
+                {{ block.titleGeneral }}
+            </h2>
 
-        <div
-            v-if="block.sectionSummary"
-            class="section-summary"
-            v-html="block.sectionSummary"
-        />
+            <div
+                v-if="block.sectionSummary"
+                class="section-summary"
+                v-html="block.sectionSummary"
+            />
+        </div>
 
         <flexible-media-gallery-new-lightbox
             v-if="showLightboxModal"
@@ -124,18 +126,18 @@ export default {
     margin: 0 auto;
     background-color: var(--color-theme, var(--color-white));
 
+    .section-header {
+        margin-bottom: var(--space-xl);
+    }
+
     .section-title {
         @include step-3;
-        margin-bottom: var(--space-xl);
+        margin-bottom: var(--space-m);
         color: var(--color-primary-blue-03);
     }
 
     .section-summary {
         @include step-0;
-        margin-bottom: var(--space-xl);
-        // ::v-deep p {
-        //     margin: 0;
-        // }
     }
 
     .thumbnails {