+++ src/lib-components/Flexible/MediaGallery/NewLightbox.vue
@@ -50,13 +50,14 @@
                     <svg-icon-molecule-bullet />
                 </button>
             </div>
+
             <h4
                 class="media-object-title"
-                v-text="images[selectionIndex].title"
+                v-text="captionTitle[selectionIndex]"
             />
             <p
                 class="media-object-caption"
-                v-text="images[selectionIndex].alt"
+                v-text="captionText[selectionIndex]"
             />
         </div>
     </div>
@@ -104,6 +105,12 @@ export default {
     computed: {
         images() {
             return this.block.map((item) => item.image[0])
+        },
+        captionTitle() {
+            return this.block.map((item) => item.captionTitle)
+        },
+        captionText() {
+            return this.block.map((item) => item.captionText)
         },
     },
     mounted() {