+++ src/lib-components/Flexible/MediaWithText.vue
@@ -108,7 +108,7 @@ export default {
     }
 
     .section-title {
-        @include step-4;
+        @include step-3;
         color: var(--color-primary-blue-03);
         margin-bottom: var(--space-m);
     }